feat: move logic to report store service
