+++ projects/ontimize-web-ngx-report/src/lib/components/report/o-report-detail/o-report-detail.component.html
@@ -1,4 +1,4 @@
-<o-form #form attr="report" service-type="report" entity="getReport" keys="UUID" header-actions="R;I;U;D" show-header="yes"
+<o-form #form attr="report" service-type="reportstore" entity="getReport" keys="UUID" header-actions="R;I;U;D" show-header="yes"
   show-header-navigation="yes" fxLayout="column" (onDataLoaded)="onDataLoaded($event)">
   <o-column attr="row1" title="REPORT_DATA" icon="description" layout-gap="14px" fxFill>
     <o-text-input #id attr="UUID" label="REPORT_ID" enabled="no" fxFlex="100"></o-text-input>
@@ -12,7 +12,7 @@
 
   <div *ngIf="hasParams">
     <o-column #params title="PARAMETERS" icon="info" class="vertical-margin-10" layout-align="start stretch" appearance="outline">
-      <o-form #paramForm attr="reportParams" service-type="report" entity="fillReport" keys="id" header-mode="none" show-header="no"
+      <o-form #paramForm attr="reportParams" service-type="reportstore" entity="fillReport" keys="id" header-mode="none" show-header="no"
         layout-direction="column" query-on-init="no" confirm-exit="no">
         <div fxLayout="column wrap" fxLayoutGap="14px" fxLayoutAlign="start start" fxFlex>
           <ng-container *ngFor="let elem of parameters; let i = index" [ngSwitch]="elem.valueClass">
