+++ projects/ontimize-web-ngx-report/src/lib/components/report/o-report-home/o-report-home.component.html
@@ -1,6 +1,6 @@
-<o-form-layout-manager attr="report" title="{{'REPORTS' | oTranslate }}" separator=" " mode="tab" label-columns="NAME">
+<o-form-layout-manager attr="report-layout-mgn" title="{{'REPORTS' | oTranslate }}" separator=" " mode="tab" label-columns="NAME">
   <div fxLayout="column" fxLayoutAlign="start start" fxFill>
-    <o-table #table fxFlex attr="report" service-type="report" entity="listReports" keys="UUID"
+    <o-table #table fxFlex attr="report" service-type="reportstore" entity="listReports" keys="UUID"
       columns="UUID;NAME;DESCRIPTION;REPORT_TYPE;MAIN_REPORT_FILENAME"
       visible-columns="NAME;DESCRIPTION;REPORT_TYPE" query-rows="20" show-report-on-demand-option="no">
       <o-table-column attr="NAME" title="{{'REPORT_NAME' | oTranslate }}" orderable="yes" searchable="yes" class="o-table-column-centered"></o-table-column>
