+++ projects/ontimize-web-ngx-report/src/lib/services/o-report-store.service.ts
@@ -1,43 +1,113 @@
 import { Injectable, Injector } from "@angular/core";
 import { MatDialog } from "@angular/material";
-import { DialogService } from "ontimize-web-ngx";
+import { DialogService, Observable, Util } from "ontimize-web-ngx";
 import { OReportViewerComponent } from "../components/report/o-report-viewer/o-report-viewer.component";
 import { Utils } from "../util/utils";
 import { OReportService } from "./o-report.service";
+import { HttpEventType, HttpRequest } from "@angular/common/http";
+import { share } from 'rxjs/operators';
+import { OReportStoreParam, OReportStoreParamValue } from "../types/report-store-param.type";
+import { OFilterParameter } from "../types/filter-parameter.type";
 
-@Injectable()
-export class OReportStoreService {
-  private reportService: OReportService;
+@Injectable({ providedIn: 'root' })
+export class OReportStoreService extends OReportService {
   protected dialogService: DialogService;
   protected dialog: MatDialog
+
   constructor(
-    private injector: Injector
+    protected injector: Injector
   ) {
+    super(injector);
     this.dialogService = this.injector.get<DialogService>(DialogService);
     this.dialog = this.injector.get<MatDialog>(MatDialog);
-    this.reportService = this.injector.get<OReportService>(OReportService);
   }
 
-  openFillReport(reportId: string, parametersValues: object, filter: object) {
-    this.reportService.configureService(this.reportService.getDefaultServiceConfiguration());
-    this.reportService.configureAdapter();
-    let kv = { 'UUID': reportId };
-    this.reportService.query(kv, null, 'getReport', {}).subscribe(
+  public query(kv?: Object, _av?: Array<string>, entity?: string, _sqltypes?: Object): Observable<any> {
+    const identifier = kv['UUID'];
+    let url = '';
+    if (Object.keys(kv).length === 0) {
+      url = `${this.urlBase}/reportstore/${entity}`;
+    } else {
+      url = `${this.urlBase}/reportstore/${entity}/` + identifier;
+    }
+
+    return this.doRequest({
+      method: 'GET',
+      url: url
+    });
+  }
+
+
+  public advancedQuery(_kv?: Object, _av?: Array<string>, entity?: string, _sqltypes?: Object, offset?: number, _pagesize?: number, _orderby?: Array<Object>): Observable<any> {
+    offset = (Util.isDefined(offset)) ? offset : this.offset;
+
+    // Calculate page
+    let page = 0;
+    if (Util.isDefined(offset)) {
+      page = Math.trunc(offset / 10) + 1;
+    }
+
+    let url = this.urlBase + '/reportstore/' + entity + '/?format=json' + '&page=' + page;
+
+    return this.doRequest({
+      method: 'GET',
+      url: url
+    });
+  }
+
+  public delete(kv?: Object, _entity?: string, _sqltypes?: Object): Observable<any> {
+    const identifier = kv.valueOf()[Object.keys(kv)[0]];
+    let url = `${this.urlBase}/reportstore/removeReport/` + identifier;
+
+    return this.doRequest({
+      method: 'DELETE',
+      url: url
+    });
+  }
+
+  public update(kv?: Object, av?: Array<string>, _entity?: string, _sqltypes?: Object): Observable<any> {
+    const identifier = kv.valueOf()[Object.keys(kv)[0]];
+    let url = `${this.urlBase}/reportstore/updateReport/` + identifier;
+
+    return this.doRequest({
+      method: 'PUT',
+      url: url,
+      body: av
+    });
+  }
+
+  public fillReport(uuid: string, reportStoreParam: OReportStoreParam, entity?: string, _sqltypes?: Object): Observable<any> {
+    let body = JSON.stringify(reportStoreParam);
+    let url = `${this.urlBase}/reportstore/${entity}/` + uuid;
+
+    return this.doRequest({
+      method: 'POST',
+      url: url,
+      body: body
+    });
+  }
+
+  openFillReport(uuid: string, parametersValues: Array<OReportStoreParamValue> = [], filter: OFilterParameter = {filter:{}}) {
+    this.configureService(this.getDefaultServiceConfiguration());
+    this.configureAdapter();
+    let kv = { 'UUID': uuid };
+    this.query(kv, null, 'getReport', {}).subscribe(
       res => {
         if (res && res.data.length && res.code === 0) {
-          let parameters = res.data[0].PARAMETERS;
           let name = res.data[0].NAME;
-          let av = [];
-          if (parameters.length > 0) {
-            av = [reportId];
-            let values = Object.values(parametersValues);
-            for (let value of values) {
-              av.push(value);
-            }
-          } else {
-            av = [reportId];
+
+          // let parameters = res.data[0].PARAMETERS;
+          // TODO check if all passed parameters are correct
+
+          const reportStoreParam: OReportStoreParam = {
+            filters: filter,
+            parameters: parametersValues
           }
-          const data = { 'params': av, 'filter': filter, 'name': name };
+          const data = {
+            'id': uuid,
+            'name': name,
+            'param': reportStoreParam
+          };
           Utils.openModalVisor(this.dialog, OReportViewerComponent, data)
 
         }
@@ -52,4 +122,57 @@ export class OReportStoreService {
     );
 
   }
+
+  upload(files: any[], entity: string, data?: object): Observable<any> {
+    const dataObservable = new Observable(observer => {
+
+      let url = `${this.urlBase}/reportstore/${entity}`;
+
+      const toUpload: any = new FormData();
+      files.forEach(item => {
+        item.prepareToUpload();
+        item.isUploading = true;
+        toUpload.append('name', item.name);
+        toUpload.append('file', item.file);
+      });
+
+      if (data) {
+        toUpload.append('data', JSON.stringify(data));
+      }
+
+      const request = new HttpRequest('POST', url, toUpload, {
+        headers: this.buildHeadersReport(),
+        reportProgress: true
+      });
+
+      this.httpClient.request(request).subscribe(resp => {
+        if (HttpEventType.UploadProgress === resp.type) {
+          // Upload progress event received
+          const progressData = {
+            loaded: resp.loaded,
+            total: resp.total
+          };
+          observer.next(progressData);
+        } else if (HttpEventType.Response === resp.type) {
+          // Full response received
+          if (resp.body) {
+            this.bodyCode(resp, observer);
+          } else {
+            observer.next(resp.body);
+          }
+        }
+      }, error => {
+        console.error(error);
+        if (error.status === 401) {
+          this.authService.logout();
+        } else {
+          observer.error(error);
+        }
+      },
+        () => observer.complete());
+    });
+    return dataObservable.pipe(share());
+  }
+
+
 }
