+++ projects/ontimize-web-ngx-report/src/lib/services/o-report.service.ts
@@ -1,9 +1,7 @@
 import { Injectable, Injector } from "@angular/core";
 import { OReportParam } from "../types/report-param.type";
 import { Observable, OErrorDialogManager, OntimizeEEService, ServiceRequestParam, ServiceResponse, Util } from 'ontimize-web-ngx';
-import { HttpEventType, HttpHeaders, HttpRequest } from "@angular/common/http";
-import { share } from 'rxjs/operators';
-import { OReportStoreParam } from "../types/report-store-param.type";
+import { HttpHeaders } from "@angular/common/http";
 
 
 @Injectable({ providedIn: 'root' })
@@ -117,56 +115,6 @@ export class OReportService extends OntimizeEEService {
     this.oErrorDialogManager.openErrorDialog(error);
   }
 
-  upload(files: any[], entity: string, data?: object): Observable<any> {
-    const dataObservable = new Observable(observer => {
-
-      let url = `${this.urlBase}/reportstore/${entity}`;
-
-      const toUpload: any = new FormData();
-      files.forEach(item => {
-        item.prepareToUpload();
-        item.isUploading = true;
-        toUpload.append('name', item.name);
-        toUpload.append('file', item.file);
-      });
-
-      if (data) {
-        toUpload.append('data', JSON.stringify(data));
-      }
-
-      const request = new HttpRequest('POST', url, toUpload, {
-        headers: this.buildHeadersReport(),
-        reportProgress: true
-      });
-
-      this.httpClient.request(request).subscribe(resp => {
-        if (HttpEventType.UploadProgress === resp.type) {
-          // Upload progress event received
-          const progressData = {
-            loaded: resp.loaded,
-            total: resp.total
-          };
-          observer.next(progressData);
-        } else if (HttpEventType.Response === resp.type) {
-          // Full response received
-          if (resp.body) {
-            this.bodyCode(resp, observer);
-          } else {
-            observer.next(resp.body);
-          }
-        }
-      }, error => {
-        console.error(error);
-        if (error.status === 401) {
-          this.authService.logout();
-        } else {
-          observer.error(error);
-        }
-      },
-        () => observer.complete());
-    });
-    return dataObservable.pipe(share());
-  }
   protected bodyCode(resp, observer) {
     if (resp.body['code'] === 3) {
       this.authService.logout();
@@ -179,71 +127,6 @@ export class OReportService extends OntimizeEEService {
       // Unknow state -> error
       observer.error('Service unavailable');
     }
-  }
-
-
-  public advancedQuery(_kv?: Object, _av?: Array<string>, entity?: string, _sqltypes?: Object, offset?: number, _pagesize?: number, _orderby?: Array<Object>): Observable<any> {
-    offset = (Util.isDefined(offset)) ? offset : this.offset;
-
-    // Calculate page
-    let page = 0;
-    if (Util.isDefined(offset)) {
-      page = Math.trunc(offset / 10) + 1;
-    }
-
-    let url = this.urlBase + '/reportstore/' + entity + '/?format=json' + '&page=' + page;
-
-    return this.doRequest({
-      method: 'GET',
-      url: url
-    });
-  }
-
-  public query(kv?: Object, _av?: Array<string>, entity?: string, _sqltypes?: Object): Observable<any> {
-    const identifier = kv['UUID'];
-    let url = '';
-    if (Object.keys(kv).length === 0) {
-      url = `${this.urlBase}/reportstore/${entity}`;
-    } else {
-      url = `${this.urlBase}/reportstore/${entity}/` + identifier;
-    }
-
-    return this.doRequest({
-      method: 'GET',
-      url: url
-    });
-  }
-
-  public fillReport(uuid: string, reportStoreParam: OReportStoreParam, entity?: string, _sqltypes?: Object): Observable<any> {
-    let body = JSON.stringify(reportStoreParam);
-    let url = `${this.urlBase}/reportstore/${entity}/` + uuid;
-
-    return this.doRequest({
-      method: 'POST',
-      url: url,
-      body: body
-    });
-  }
-
-  public delete(kv?: Object, _entity?: string, _sqltypes?: Object): Observable<any> {
-    const identifier = kv.valueOf()[Object.keys(kv)[0]];
-    let url = `${this.urlBase}/reportstore/removeReport/` + identifier;
-
-    return this.doRequest({
-      method: 'DELETE',
-      url: url
-    });
-  }
-
-  public update(kv?: Object, av?: Array<string>, _entity?: string, _sqltypes?: Object): Observable<any> {
-    const identifier = kv.valueOf()[Object.keys(kv)[0]];
-    let url = `${this.urlBase}/reportstore/updateReport/` + identifier;
-
-    return this.doRequest({
-      method: 'PUT',
-      url: url,
-      body: av
-    });
   }
 
   protected buildHeadersReport(): HttpHeaders {
