+++ projects/ontimize-web-ngx-report/src/lib/types/filter-parameter.type.ts
@@ -1,7 +1,9 @@
+import { BasicExpression } from "ontimize-web-ngx"
+
 export type OFilterParameter = {
-  columns: Array<Object>,
-  sqltypes: Map<Object, Object>,
-  filter: Map<Object, Object>,
-  offset: number,
-  pageSize: number
+  filter: object | BasicExpression,
+  columns?: Array<string>,
+  sqltypes?: object,
+  offset?: number,
+  pageSize?: number
 }
